[Feature] Adding unit tests for indexes (#31)
* [Refactor] moved util code to separate header files

* [FEAT] added unit tests

* Added detailed instructions on integrating new index to readme

* minor refactor to define

* minor refactor to header defs
