+++ tests/test_btree/btree_tests.cpp
@@ -1,0 +1,31 @@
+#include "bliss_index_tests.h"
+
+class BTreeTest : public BlissIndexTest {};
+
+TEST_F(BTreeTest, TestBTree_Sorted) {
+    index.reset(new bliss::BlissBTreeIndex<key_type, key_type>());
+    std::vector<key_type> data;
+    GenerateData(data, num_keys);
+
+    auto insert_start = data.begin();
+    auto insert_end = data.end();
+    executor::execute_inserts(*index, insert_start, insert_end);
+
+    for (auto key : data) {
+        EXPECT_TRUE(index->get(key));
+    }
+}
+
+TEST_F(BTreeTest, TestBTree_Random) {
+    index.reset(new bliss::BlissBTreeIndex<key_type, key_type>());
+    std::vector<key_type> data;
+    GenerateData(data, num_keys, false);
+
+    auto insert_start = data.begin();
+    auto insert_end = data.end();
+    executor::execute_inserts(*index, insert_start, insert_end);
+
+    for (auto key : data) {
+        EXPECT_TRUE(index->get(key));
+    }
+}
