+++ tests/test_lipp/lipp_tests.cpp
@@ -1,0 +1,31 @@
+#include "bliss_index_tests.h"
+
+class LippTest : public BlissIndexTest {};
+
+TEST_F(LippTest, TestLipp_Sorted) {
+    index.reset(new bliss::BlissLippIndex<key_type, key_type>());
+    std::vector<key_type> data;
+    GenerateData(data, num_keys);
+
+    auto insert_start = data.begin();
+    auto insert_end = data.end();
+    executor::execute_inserts(*index, insert_start, insert_end);
+
+    for (auto key : data) {
+        EXPECT_TRUE(index->get(key));
+    }
+}
+
+TEST_F(LippTest, TestLipp_Random) {
+    index.reset(new bliss::BlissAlexIndex<key_type, key_type>());
+    std::vector<key_type> data;
+    GenerateData(data, num_keys, false);
+
+    auto insert_start = data.begin();
+    auto insert_end = data.end();
+    executor::execute_inserts(*index, insert_start, insert_end);
+
+    for (auto key : data) {
+        EXPECT_TRUE(index->get(key));
+    }
+}
